Cleanup state functions and data formats
This cleanup is the first stage in converting to not just using method
names but also arities and so become properly erlang compliant.

We define quite a few data functions in the '-core' modules with
pre-processed data, many more than is actually necessary. While this
means more work in this stage it means less when we actually make a
flavor instance. We will see if we continue in this way or trim.